[level 1] Title: 정수 제곱근 판별, Time: 11.40 ms, Memory: 87.1 MB -BaekjoonHub

+++ 프로그래머스/1/12934. 정수 제곱근 판별/정수 제곱근 판별.java
@@ -1,11 +1,11 @@
 class Solution {
-    public double solution(long n) {
+    public long solution(long n) {
         if(n == 1) 
             return 4;
         double sqrt = Math.sqrt(n);
         for(int i=1; i<n/2; i++){
             if(sqrt == i){
-                return Math.pow(i+1, 2);
+                return (long)Math.pow(i+1, 2);
             }
         }
         return -1;
